Refactorization in the spirit of the model evaluator
The Model Evaluator is a purely abstract class and as such should not hold any data. Up to this point, this was not entirely true for the Ginla model evalator as the kinetic energy operator was stored in the class explicitly, and shared amongst all components that needed it, e.g., the function evaluation, the Jacobian, adn the preconditioner. Its entries were adapted on an as-needed basis.
Now, for each instance that needs the KEO, a separate copy is hosted. When requested by evalModel(), the entries are updated.

Besides that, an explicit KEO preconditioner class has been added that essientially is a KEO unless the KEO is (near) singular; then, a regularization constant is added to the diagonal.