Major refactor of the code base
It had turned out in the past that keeping too many virtual classes around to ensure compatibility across the implementations (FDM, FVM, FVM Tpetra) limited the agility in programming as for every change to those virtual classes, all classes had to be adapted.
The new rationale is to keep the implementations separate, although this results in code duplication. This drawback however is considered less harmful than the delay that is caused by enforcing too much compatibility.