towards a truly stateless model evaluator
This is a rather large commit, and mainly gets rid of
storing parameters anywhere (except for caching). The commit
also includes a move of parameter specification in the
potential classes themselves. This makes it possible
to derive custom potentials in classes derived from
the virtuals to hook up with this same code.
i#	modified:   Ginla_ModelEvaluator.hpp